+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KQFFVBT4Y45N4VJ3E7HV8PTS"
+        "01KQGZE4G5JP26ZMKPB1VXFACE"
     )
     r"""The unique identifier of the data source"""