+++ packages/orq-rc/src/orq_ai_sdk/models/getpromptversionop.py
@@ -315,6 +315,7 @@ r"""Controls the verbosity of the model output."""
 
 GetPromptVersionThinkingLevel = Literal[
     "low",
+    "medium",
     "high",
 ]
 r"""The level of thinking to use for the model. Only supported by `Google AI`"""