+++ packages/orq-rc/RELEASES.md
@@ -7459,3 +7459,13 @@ Based on:
 - [python v4.9.0-rc.13] packages/orq-rc
 ### Releases
 - [PyPI v4.9.0-rc.13] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.13 - packages/orq-rc
+
+## 2026-05-01 05:18:49
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.10 (2.881.16) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.9.0-rc.14] packages/orq-rc
+### Releases
+- [PyPI v4.9.0-rc.14] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.14 - packages/orq-rc